backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi-cors>=0.0.6",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[standard]>=0.116.1",
    "openai>=1.107.2",
    "pocketflow>=0.0.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
